# [codegen/node] Implement support for lazy-loaded Node modules