--- a/pkg/codegen/testing/test/testdata/azure-native-nested-types/nodejs/types/index.ts
+++ b/pkg/codegen/testing/test/testdata/azure-native-nested-types/nodejs/types/index.ts
@@ -3,7 +3,7 @@
 
 import * as utilities from "../utilities";
 
-// Export sub-modules (modContext.genIndex):
+// Export sub-modules:
 import * as input from "./input";
 import * as output from "./output";