… CHANGELOG_PENDING.md
… pkg/codegen/nodejs/gen_fileinfo.go
… pkg/codegen/nodejs/gen_lazyloads.go
… pkg/codegen/nodejs/gen_lazyloads_test.go
… pkg/codegen/nodejs/importer.go
… pkg/codegen/nodejs/submod_export.go
… pkg/codegen/nodejs/submod_export_test.go
… pkg/codegen/nodejs/utilities.go
… pkg/codegen/testing/test/testdata/azure-native-nested-types/nodejs/documentdb/index.ts
… pkg/codegen/testing/test/testdata/azure-native-nested-types/nodejs/documentdb/sqlResourceSqlContainer.ts
… pkg/codegen/testing/test/testdata/azure-native-nested-types/nodejs/index.ts
… pkg/codegen/testing/test/testdata/azure-native-nested-types/nodejs/types/input.ts
… pkg/codegen/testing/test/testdata/azure-native-nested-types/nodejs/types/output.ts
… pkg/codegen/testing/test/testdata/azure-native-nested-types/nodejs/utilities.ts
… pkg/codegen/testing/test/testdata/azure-native-nested-types/schema.json
… pkg/codegen/testing/test/testdata/cyclic-types/nodejs/index.ts
… pkg/codegen/testing/test/testdata/cyclic-types/nodejs/types/index.ts
… pkg/codegen/testing/test/testdata/cyclic-types/nodejs/types/input.ts
… pkg/codegen/testing/test/testdata/cyclic-types/nodejs/types/output.ts
… pkg/codegen/testing/test/testdata/cyclic-types/nodejs/utilities.ts
… pkg/codegen/testing/test/testdata/dash-named-schema/nodejs/index.ts
… pkg/codegen/testing/test/testdata/dash-named-schema/nodejs/submodule1/index.ts
… pkg/codegen/testing/test/testdata/dash-named-schema/nodejs/submodule1/moduleResource.ts
… pkg/codegen/testing/test/testdata/dash-named-schema/nodejs/types/index.ts
… pkg/codegen/testing/test/testdata/dash-named-schema/nodejs/types/input.ts
… pkg/codegen/testing/test/testdata/dash-named-schema/nodejs/types/output.ts
… pkg/codegen/testing/test/testdata/dash-named-schema/nodejs/utilities.ts
… pkg/codegen/testing/test/testdata/dashed-import-schema/nodejs/index.ts
… pkg/codegen/testing/test/testdata/dashed-import-schema/nodejs/tree/index.ts
… pkg/codegen/testing/test/testdata/dashed-import-schema/nodejs/tree/v1/index.ts