<!--- 
Thanks so much for your contribution! If this is your first time contributing, please ensure that you have read the [CONTRIBUTING](https://github.com/pulumi/pulumi/blob/master/CONTRIBUTING.md) documentation.
-->

# Description

<!--- Please include a summary of the change and which issue is fixed. Please also include relevant motivation and context. -->

This PR implements opt-in support for generating lazy load code for Node modules. The changes should be entirely non-breaking for consumers of the providers.

https://github.com/pulumi/pulumi/issues/7653 for more history and motivation

In addition to PR checks, this has been tested manually on azure-native. It compiles and delivers speedups in node startup time: from 7.6s to 1.6s when importing "@pulumi/azure-native" and from 800ms to 450ms when importing "@pulumi/azure-native/resources" and "@pulumi/azure-native/storage".

How this works:

- modules that define functions  and resources are loaded lazily 
- generated code may use `import type ..` form when importing enums ("useTypeOnlyReferences" opt-in flag in schema)

`pulumi.runtime.registerResourceModule` side-effects are not affected and happen as before.

Our prior approach on 7653 went for lazily importing module-re-exports; this would still be preferable but we run into some corner cases where we can't seem to properly make TypeScript re-export the module as a "namespace", leading to a class of programs that can break.  This is explored in some detail in https://github.com/pulumi/pulumi-azure-native/pull/1944

Further work: we should be able to do a lot better here with subsequent work by making sure we emit more precise imports, this will optimize not only node time but also TS compilation time - this is tracked in https://github.com/pulumi/pulumi/issues/10442 - but recommend adopting this for now.

Fixes #7653 

## Checklist

<!--- Please provide details if the checkbox below is to be left unchecked. -->
- [ ] I have added tests that prove my fix is effective or that my feature works
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [ ] I have updated the [CHANGELOG-PENDING](https://github.com/pulumi/pulumi/blob/master/CHANGELOG_PENDING.md) file with my change
<!--
If the change(s) in this PR is a modification of an existing call to the Pulumi Service,
then the service should honor older versions of the CLI where this change would not exist.
You must then bump the API version in /pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi Service API version
  <!-- @Pulumi employees: If yes, you must submit corresponding changes in the service repo. -->