@t0yv0 2022-08-31
I've found and fixed some issues by trying it on pulumi-azure-native (not caught by codegen tests? need to look into this).

This helps! Test: `pulumi new azure-javascript` (for some reason this returns the "wrong" style import).

# After

```
Benchmark 1: node index.js
  Time (mean ± σ):      4.476 s ±  0.191 s    [User: 3.973 s, System: 0.959 s]
  Range (min … max):    4.257 s …  4.859 s    10 runs
```

# Before

```
Benchmark 1: node index.js
  Time (mean ± σ):      7.944 s ±  0.399 s    [User: 7.208 s, System: 1.670 s]
  Range (min … max):    7.304 s …  8.521 s    10 runs
 
  Warning: Ignoring non-zero exit code.
 
```