@t0yv0 2022-08-31
Tracing which modules get loaded, it's the function modules mostly. This is very promising. IF we lazy-load them also we get a real win here.

```
NODE_DEBUG=module node index.js 2>module-load.log
cat module-load.log | grep "load[ ]" | grep -v REQUEST | grep azure-native/sdk/nodejs/bin | grep -v get  | grep -v list | grep -v index
```