@t0yv0 2022-08-31
I'm starting to think this approach has an advantage over lazy top-level submodules only - this optimizes imports of `@pulumi/azure-native/resources` as well as top-level `@pulumi/azure-native`