@t0yv0 2022-08-31
Reminder to self: this PR still needs to address the fate of resource registrations. AFAIK these must be eagerly registered when importing any of the modules as before. The lazy loading may defer the registrations too late.