@t0yv0 2022-08-31
The lazy-load of functions had impact! Now on azure-native:

```
Benchmark 1: node index.js
  Time (mean ± σ):      1.652 s ±  0.058 s    [User: 1.697 s, System: 0.233 s]
  Range (min … max):    1.593 s …  1.773 s    10 runs
 
  Warning: Ignoring non-zero exit code.

```