@t0yv0 2022-08-31
Now measure the example rewritten for recommended imports "azure-native/storage" and "azure-native/resources".

```
BEFORE:

Benchmark 1: node index.js
  Time (mean ± σ):     870.1 ms ±  20.2 ms    [User: 879.4 ms, System: 133.5 ms]
  Range (min … max):   850.4 ms … 917.1 ms    10 runs

AFTER:

Benchmark 1: node index.js
  Time (mean ± σ):     871.9 ms ±  39.1 ms    [User: 893.2 ms, System: 126.5 ms]
  Range (min … max):   846.0 ms … 975.6 ms    10 runs
```

It's a wash because it manages to load all enums from all modules, which dominates. This needs addressing somehow to be a real benefit to users. Tinkering further.