@t0yv0 2022-08-31
Ok good, with the latest tweak to avoid importing types/index.js that gets us all enum modules imported, we reclaim the last milliseconds.

Code:

```
"use strict";

const pulumi = require("@pulumi/pulumi");
const resources = require("@pulumi/azure-native/resources");
const storage = require("@pulumi/azure-native/storage/storageAccount");
```

After:

```
Benchmark 1: node index.js
  Time (mean ± σ):     436.7 ms ±  34.8 ms    [User: 450.0 ms, System: 51.9 ms]
  Range (min … max):   394.0 ms … 496.1 ms    10 runs

  Warning: Ignoring non-zero exit code.
```

Before:

```
Benchmark 1: node index.js
  Time (mean ± σ):     818.4 ms ±  21.5 ms    [User: 823.3 ms, System: 124.5 ms]
  Range (min … max):   798.2 ms … 870.2 ms    10 runs

  Warning: Ignoring non-zero exit code.
```

This is clearly better and we only import 50 modules from azure-native now (mostly the index.js ones), out of 650 total modules imported. I think this is good enough.