@t0yv0 2022-08-31
> Reminder to self: this PR still needs to address the fate of resource registrations. AFAIK these must be eagerly registered when importing any of the modules as before. The lazy loading may defer the registrations too late.

Ah nevermind! We are lucky here. The calls to registerResourceModule are all sitting in index.ts modules files which are not affected by this change. So they will happen exactly as before. As to enum files, they never call registerResourceModule. So this concern is irrelevant for the PR.