@t0yv0 2022-09-01
Downstream checks actually passed here: https://github.com/pulumi/pulumi/pull/10576 after a fixup to the GHA definition.